ENGR00172490: FIX: PMIC registers may not be accessible in u-boot via SPI
The system PMIC registers may not be accessible in u-boot via SPI if
function pmic_reg() is called in the latter part of boot up process in u-boot.

It is because the imx_spi_slave structure is allocated from malloc() in
the spi_setup_slave() function. However, this structure is not completely
initialized, which may result in using a dirty control register value
at CSPI during transfer.

memset() the imx_spi_slave structure after malloc() can resolve this problem

Please refer to CT39243849.

Signed-off-by: Robby Cai <R63905@freescale.com>