ENGR00217764 MX6 Secure Boot : Fix NAND BOOT Failure due to secure patch
With the secure boot patch. MX6 NAND Boot is not functional. The root
cause is that, the original secure boot patch fills "0xFF' to spacing
regions, due to a issue in ROM code, read pages of all "0xff" will be
treated as a critical error. Thus prevent the U-Boot from booting
normally.

The fix adjust image copy size in IVT so that when secure boot is not
enabled, no unuseful data is copied by ROM code. Also the secure boot
option is default disabled. The end user won't enable it unless they
know what they are doing.

These prevent the ROM code from copied pages of "0xff" data, and fix the
issue.

Signed-off-by: Eric Sun <jian.sun@freescale.com>