cm-fx6: add support for on-board SSD
Add support for the on-board SanDisk SSD.
Linux SATA and SSD initialization procedure fails after the SSD has been
initialized in U-Boot.
To workaround the problem (and for compatibility with all Linux kernel
versions) we power the SSD off prior to booting into Linux.
To be less intrusive we use the show_boot_progress() function as a
callback for SSD shut down which is a bit hackish.

Signed-off-by: Igor Grinberg <grinberg@compulab.co.il>